```markdown Convert form number fields from `int` to `int64` for consistent type safety and support of larger sequence values
- Changed `formNumberLastYear`, `formNumberLastMonth`, and `formNumberLastSequence` from `int` to `int64` in form number management logic
  - (AI deducted reason: Ensures compatibility with larger sequence values that may exceed `int` range, preventing overflow in edge cases during monthly resets or long-term form numbering)

- Aligns field types with database upsert query results which return sequence values as `int64`
  - (AI deducted reason: Matches data type consistency across the module, avoiding implicit type coercion or casting errors during form number generation)

- Supports future expansion of sequence values (e.g., for high-volume forms or long-term systems) without requiring structural changes
  - (AI deducted reason: Future-proofing the module to handle larger numbers, especially when sequences are reset monthly and grow over time)
```

+++ module/form_number_management/form_number_management.go
@@ -80,9 +80,9 @@ func (fnm *DxmFormNumberManagement) Generate(nameid string, timezone string) (st
 	formNumberType := rr["type"].(string)
 	formNumberTemplate := rr["template"].(string)
 	formNumberPrefix := rr["prefix"].(string)
-	formNumberLastYear := rr["last_year"].(int)
-	formNumberLastMonth := rr["last_month"].(int)
-	formNumberLastSequence := rr["last_sequence"].(int)
+	formNumberLastYear := rr["last_year"].(int64)
+	formNumberLastMonth := rr["last_month"].(int64)
+	formNumberLastSequence := rr["last_sequence"].(int64)
 
 	// Format form number
 	formNumber := ""
